added custom list delete fuction and more

+++ app.js
@@ -3,6 +3,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 
 const app = express();
 
@@ -55,43 +56,65 @@ app.get("/", async function(req, res){
 });
 
 app.get("/:customListName", async function(req, res){
-    const customListName =  req.params.customListName;
+
+    const customListName =  _.capitalize(req.params.customListName);
 
     const foundList = await List.findOne({name: customListName})
 
-        if(!foundList){
-            const list = new List({
-                name: customListName,
-                items: defaulItem
-            });
-            list.save();
-            res.redirect("/" + customListName);
-        }
-        else{
-            res.render("list", {listTitle: foundList.name, newListItems: foundList.items});
-        }
+    if(!foundList){
+        const list = new List({
+            name: customListName,
+            items: defaulItem
+        });
+        list.save();
+        res.redirect("/" + customListName);
+    }
+    else{
+        res.render("list", {listTitle: foundList.name, newListItems: foundList.items});
+    }
     
 });
 
-app.post("/", function(req, res){
+app.post("/", async function(req, res){
 
     const itemName = req.body.newItem;
+    const listName = req.body.list;
 
     const item = new Item({
         name: itemName
     });
 
-    item.save();
-    res.redirect("/");
+    if(listName === "Today"){
+        item.save();
+        res.redirect("/");
+    }
+    else{
+        const foundList = await List.findOne({name: listName});
+        foundList.items.push(item);
+        foundList.save();
+        res.redirect("/" + listName);
+    }
+
+    
 });
 
 app.post("/delete", async function(req, res){
 
     const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
 
-    const deletedItem = await Item.findByIdAndRemove(checkedItemId);
+    if(listName === "Today"){
+        const deletedItem = await Item.findByIdAndRemove(checkedItemId);
+        res.redirect("/");
+    }
+    else{
+        const foundList = await List.findOneAndUpdate({name: listName}, {$pull: {items: {_id: checkedItemId}}});
+        res.redirect("/" + listName);
+    }
 
-    res.redirect("/");
+    
+
+   
 
 });
